소스 최대 사용 가격 제한 추가 소스 드래그 시작 및 드래그 종료 시 소스 바르도록 수정

+++ Assets/Scripts/UI/ToggleWindow.cs
@@ -21,7 +21,7 @@ public class ToggleWindow : MonoBehaviour
         {
             if (SaveLoadManager.Data.unlocks.TryGetValue(ing.ingredientID, out bool val))
             {
-                var tog = Instantiate(prefab, contents[ing.type - 1]);
+                var tog = Instantiate(prefab, contents[(int)ing.type - 1]);
                 tog.Init(ing.ingredientID, val);
             }
         }
